fixed several bugs in lexer and extracted content of test class for lexer contract

+++ javabite-common/src/main/java/swp_compiler_ss13/javabite/token/NumTokenJb.java
@@ -15,6 +15,6 @@ public class NumTokenJb extends TokenJb implements NumToken {
 
 	@Override
 	public Long getLongValue() {
-		return Long.parseLong(getValue());
+		return (long)Double.parseDouble(getValue());
 	}	
 }
